LayoutTests/imported/w3c/web-platform-tests/mathml/presentation-markup/scripts/subsup-parameters-2-expected.txt


PASS Null Italic Correction
PASS NonNull Italic Correction (MathGlyphVariantRecord)
FAIL NonNull Italic Correction (GlyphAssembly) assert_approx_equals: msub base width expected 70 +/- 1 but got 10
Null Italic Correction

⫿

⫿

⫿

⫿

NonNull Italic Correction (MathGlyphVariantRecord)

⫿

⫿

⫿

⫿

NonNull Italic Correction (GlyphAssembly)

⫿

⫿

⫿

⫿
